refactor(inference): split setup.rs into focused submodules (TD-031)
Behavior-preserving decomposition of parish-inference/src/setup.rs
(~3150 LOC) into five single-responsibility submodules under setup/:

  - progress.rs    — SetupProgress trait + StdoutProgress
  - gpu_detect.rs  — GPU vendor/VRAM detection (macOS/Windows/Linux)
  - model_select.rs — VRAM-based model tier selection
  - process.rs     — OllamaProcess, VllmMlxProcess, VllmProcess,
                     RuntimeProcesses and their slot helpers
  - orchestration.rs — full Ollama setup sequence, pull/delete/warmup,
                     and the unified setup_provider_client entry point

Public API of parish-inference is unchanged: all pre-split use paths
(setup::SetupProgress, setup::RuntimeProcesses, etc.) continue to work
via re-exports in setup/mod.rs.

Tests moved alongside the code they exercise. cargo fmt + clippy clean.
cargo test: 328 passed, 2 ignored. Engine boots and plays in headless
mode using config_refactor_smoke.txt fixture.

Part of #1200 (TD-031). Does NOT close the epic.
